fix: resolve 4 issues — listing consolidation, settings persistence, metadata cache, reentrancy guard
