+++ frontend/afristore-app/src/context/WalletContext.tsx
@@ -1,12 +1,13 @@
 "use client";
 
-import { createContext, useContext, ReactNode, useMemo } from "react";
+import { createContext, useContext, ReactNode, useMemo, useCallback } from "react";
 import { useWallet, WalletState, WalletStatus } from "@/hooks/useWallet";
 import {
   useMagicWallet,
   MagicWalletState,
   MagicWalletStatus,
 } from "@/hooks/useMagicWallet";
+import { isE2eMockChain } from "@/lib/e2e-chain-mock";
 
 export type WalletType = "freighter" | "magic" | null;
 
@@ -19,10 +20,12 @@ export interface UnifiedWalletState {
   error: string | null;
   status: WalletStatus | "MAGIC_CONNECTED" | "DISCONNECTED";
   networkPassphrase: string | null;
+  network: string;
   isInstalled: boolean;
   connect: () => Promise<void>;
   disconnect: () => void;
   refresh: () => Promise<void>;
+  switchNetwork: (networkName: string) => Promise<void>;
   freighter: WalletState;
   magic: MagicWalletState;
 }
@@ -49,6 +52,34 @@ export function WalletProvider({ children }: { children: ReactNode }) {
         ? "MAGIC_CONNECTED"
         : "DISCONNECTED";
 
+  const network = freighter.networkPassphrase?.includes("Test SDF")
+    ? "testnet"
+    : freighter.networkPassphrase?.includes("Public Global")
+      ? "public"
+      : freighter.networkPassphrase
+        ? "futurenet"
+        : "public";
+
+  const switchNetwork = useCallback(async (networkName: string) => {
+    const passphrases: Record<string, string> = {
+      public: "Public Global Stellar Network ; September 2015",
+      testnet: "Test SDF Network ; September 2015",
+      futurenet: "Test SDF Future Network ; October 2022",
+    };
+
+    const passphrase = passphrases[networkName] ?? passphrases.public;
+
+    if (isE2eMockChain()) {
+      sessionStorage.setItem("e2e_network_passphrase", passphrase);
+      window.location.reload();
+      return;
+    }
+
+    console.info(
+      `Please switch Freighter to the ${networkName} network manually.`,
+    );
+  }, []);
+
   const value = useMemo(
     () => ({
       walletType,
@@ -57,16 +88,18 @@ export function WalletProvider({ children }: { children: ReactNode }) {
       isConnecting: freighter.isConnecting || magic.isConnecting,
       isWrongNetwork: freighter.isWrongNetwork,
       networkPassphrase: freighter.networkPassphrase,
+      network,
       isInstalled: freighter.isInstalled,
       error: freighter.error ?? magic.error,
       status,
       connect: freighter.connect,
       disconnect: freighter.disconnect,
       refresh: freighter.refresh,
+      switchNetwork,
       freighter,
       magic,
     }),
-    [walletType, publicKey, status, freighter, magic],
+    [walletType, publicKey, status, network, switchNetwork, freighter, magic],
   );
 
   return (
